converted app.js to ts; put ssh redirect in index
